+++ cmake.config/iwyu/mapping.imp
@@ -37,6 +37,7 @@
   { include: [ '"drawscreen.h.generated.h"', private, '"nvim/drawscreen.h"', public ] },
   { include: [ '"edit.h.generated.h"', private, '"nvim/edit.h"', public ] },
   { include: [ '"eval.h.generated.h"', private, '"nvim/eval.h"', public ] },
+  { include: [ '"eval/buffer.h.generated.h"', private, '"nvim/eval/buffer.h"', public ] },
   { include: [ '"eval/decode.h.generated.h"', private, '"nvim/eval/decode.h"', public ] },
   { include: [ '"eval/encode.h.generated.h"', private, '"nvim/eval/encode.h"', public ] },
   { include: [ '"eval/executor.h.generated.h"', private, '"nvim/eval/executor.h"', public ] },
@@ -44,6 +45,7 @@
   { include: [ '"eval/typval.h.generated.h"', private, '"nvim/eval/typval.h"', public ] },
   { include: [ '"eval/userfunc.h.generated.h"', private, '"nvim/eval/userfunc.h"', public ] },
   { include: [ '"eval/vars.h.generated.h"', private, '"nvim/eval/vars.h"', public ] },
+  { include: [ '"eval/window.h.generated.h"', private, '"nvim/eval/window.h"', public ] },
   { include: [ '"event/libuv_process.h.generated.h"', private, '"nvim/event/libuv_process.h"', public ] },
   { include: [ '"event/loop.h.generated.h"', private, '"nvim/event/loop.h"', public ] },
   { include: [ '"event/multiqueue.h.generated.h"', private, '"nvim/event/multiqueue.h"', public ] },
